# software-example/frontend/src/api/rest-client/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:3001/api","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Пользователь","description":"Управление данными текущего пользователя"},{"name":"Заметки","description":"Управление заметками пользователя"},{"name":"Авторизация","description":"Методы для входа и регистрации пользователей"}],"paths":{"/notes":{"get":{"tags":["Заметки"],"summary":"Получить список заметок с пагинацией","description":"Возвращает список заметок с поддержкой текстового (BM25) или гибридного поиска (BM25 + семантический)","operationId":"getNotesList","parameters":[{"name":"searchQuery","in":"query","description":"Поисковый запрос (текстовый или семантический)","required":false,"schema":{"type":"string"}},{"name":"semantic","in":"query","description":"Использовать гибридный поиск (BM25 + семантический). Если false - обычный текстовый LIKE поиск","required":false,"schema":{"type":"boolean","default":false}},{"name":"page","in":"query","description":"Номер страницы (начиная с 0)","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","description":"Размер страницы","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"semanticLimit","in":"query","description":"Количество результатов для внутреннего поиска (только для семантического/гибридного). По умолчанию: page * size + size","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"denseWeight","in":"query","description":"Вес Dense векторного поиска (0.0 - 1.0). Только для семантического/гибридного поиска. По умолчанию: 0.4","required":false,"schema":{"type":"number","format":"float","default":0.4}},{"name":"sparseWeight","in":"query","description":"Вес Sparse векторного поиска (0.0 - 1.0). Только для семантического/гибридного поиска. По умолчанию: 0.3","required":false,"schema":{"type":"number","format":"float","default":0.3}},{"name":"textWeight","in":"query","description":"Вес текстового поиска (0.0 - 1.0). Только для семантического/гибридного поиска. По умолчанию: 0.3","required":false,"schema":{"type":"number","format":"float","default":0.3}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotesPageResponseDto"}}}}}},"post":{"tags":["Заметки"],"summary":"Создать новую заметку","description":"Создает новую заметку с обязательным заголовком","operationId":"createNote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteCreateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["Авторизация"],"summary":"Сервисный вход пользователя","description":"Проверяет секретный заголовок и на его основе регистрирует или возвращает данные пользователя","operationId":"serviceLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceLoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserIdResponseDto"}}}}}}},"/notes/{id}":{"delete":{"tags":["Заметки"],"summary":"Удалить заметку (мягкое удаление)","description":"Помечает заметку как удаленную (isDeleted = true)","operationId":"deleteNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["Заметки"],"summary":"Обновить заметку","description":"Обновляет поля заметки (partial update)","operationId":"updateNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteUpdateRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteResponseDto"}}}}}}},"/user/me":{"get":{"tags":["Пользователь"],"summary":"Получить данные профиля пользователя","description":"Возвращает информацию о текущем авторизованном пользователе","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDataResponseDto"}}}}}}}},"components":{"schemas":{"NoteCreateRequestDto":{"required":["description","title","url"],"type":"object","properties":{"title":{"maxLength":255,"minLength":0,"type":"string"},"description":{"type":"string"},"url":{"type":"string"}}},"NoteResponseDto":{"required":["createdAt","description","id","title","updatedAt","url"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"url":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ServiceLoginRequestDto":{"required":["provider","providerUserId","userName"],"type":"object","properties":{"providerUserId":{"type":"string"},"userName":{"type":"string"},"provider":{"type":"string"}}},"UserIdResponseDto":{"required":["userId"],"type":"object","properties":{"userId":{"type":"integer","format":"int64"}}},"NoteUpdateRequestDto":{"required":["description","title","url"],"type":"object","properties":{"title":{"maxLength":255,"minLength":0,"type":"string"},"description":{"type":"string"},"url":{"type":"string"}}},"UserDataResponseDto":{"required":["userId","userName"],"type":"object","properties":{"userName":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"NotesPageResponseDto":{"required":["content","isLast","page","size","totalElements","totalPages"],"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/NoteResponseDto"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"isLast":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}